Renabled "let it crash" and added a cool down phase
The master_fsm monitors over the runner_fsm and if it crashes, the
master moves onto the next test.  I removed the error handing code from
the server_logger to let it crash if the server crashes or fails to start

I also added a cooldown phase for each test to let the server calm down
from being abused.  This also theoretically allows the kernel to hang up any
sockets.